clock: add new clock module
This adds our own internal clock offering based off of clocksource.

A few benefits:
- we don't care about real-time
- clock sources utilize a trait now, can be swapped in/out for tests
- calibration is also configurable
- customized start/end TSC methods for more precise measurements
(slightly slower by few nanos)
- can convert TSC delta in one call, instead of 2x converts; faster

Sinks now expose a shared reference to their own clock, which shares
calibration with a master clock in the receiver.  Not yet clear if we
truly need a mechanism for runtime recalibration but it wouldn't take a
whole lot of work.

Signed-off-by: Toby Lawrence <toby@nuclearfurnace.com>