core: metrics auto-register themselves now
Instead of having to add facets ahead of time, we now dynamically
registers metrics as they come in if we haven't seen them before.  This
allows us to make the sink API much simpler for end users.

We've also revved up the benchmark, and the documentation, to be a
little more clear and crisp about what's going on behind the scenes.

Signed-off-by: Toby Lawrence <toby@nuclearfurnace.com>