Sink scopes can now take `&str` or `&[&str]`.
This is just a little tweak to allow users more flexibility in how they
build scopes.  Previously, if you wanted to do deep nesting in one shot,
you had to build a string with the separator character, or do multiple
`scoped` calls, which is not very ergonomic.

Now, users can pass a string or a _slice_ of strings, allowing a mix and
match of constant strings, static strings, and owned strings.

Signed-off-by: Toby Lawrence <toby@nuclearfurnace.com>