Fixed a bug with identical scopes overwriting each others metrics.
Due to the way scopes were allocating their scope IDs, duplicate scopes
would be storing metrics under different scope IDs but end up being
overwritten when a snapshop was generated: the latest version of a scope
would have their metric as the "winner".

We've now changed this so that sinks register their scope directly
instead of sending a unary message to the receiver.  This means a few
things:
- sinks get told what their scope ID is, so we can deduplicate
- we go through a RW locked mechanism (slower, but still fast)
- we have the potential to clean up scope entries when the sink drops

Synchronizing the creation of a sink on a RW lock isn't great, but in
practice, unless you're creating a lot of sinks all the same, like
probably thousands per second? then this approach should be OK.  I will
add a bench test at some point for the uncontested overhead but I'm not
super worried about it.

Dropping stale scope entries is a future optimization to help with
memory consumption but scope memory usage is likely dwarfed by a few
orders of magnitude if a user has any histograms, etc.

Signed-off-by: Toby Lawrence <toby@nuclearfurnace.com>